requirements.txt
gradio==3.50.2
Pillow==9.5.0
#PADDLEOCR
shapely
scikit-image
imgaug
pyclipper
lmdb
tqdm
numpy
visualdl
rapidfuzz
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
cython
lxml
premailer
openpyxl
attrdict
Polygon3
lanms-neo==1.0.2
#PyMuPDF<1.21.0
#VORO
torch
torchvision
numpy